delete repeated funcs
