+++ synthetic.py
@@ -35,208 +35,6 @@ class SyntheticGraphGeneration:
         self.cfg          = cfg
 
 
-    def plot_graph(self, G: nx.Graph, pos: Optional[Dict[int, Tuple[float, float]]] = None,
-                   title: str = "Graph", node_size: int = 300, node_color: str = 'skyblue', 
-                   with_labels: bool = True) -> Figure:
-        """
-        Input:
-        ----------
-        G : nx.Graph
-            The NetworkX graph to be plotted.
-        pos : Optional[Dict[int, Tuple[float, float]]], default=None
-            A dictionary of node positions, where keys are node indices and values are (x, y) coordinates.
-        title : str, default="Graph"
-            The title of the graph plot.
-        node_size : int, default=300
-            The size of the nodes in the plot.
-        node_color : str, default='skyblue'
-            The color of the nodes in the plot.
-        with_labels : bool, default=True
-            Whether or not to display labels on the nodes.
-
-        Output:
-        -------
-        Figure
-            The matplotlib figure object representing the final graph plot.
-
-        Description:
-        -----------
-            This function plots the structure of the given graph using NetworkX and matplotlib.
-            It allows for customization of the title, node size, node color, and the option to include labels.
-        """
-        plt.figure(figsize=(10, 8))
-        nx.draw(G, pos, with_labels=with_labels, node_size=node_size, node_color=node_color, 
-                edge_color='gray', font_size=10)
-        plt.title(title)
-        return plt.gcf()
-
-
-    def plot_color_graph(self, G: nx.Graph, pos: Optional[Dict[int, Tuple[float, float]]] = None, 
-                         title: str = "Graph", node_size: int = 300, with_labels: bool = True) -> Figure:
-        """
-        Input:
-        ----------
-        G : nx.Graph
-            The NetworkX graph to be plotted.
-        pos : Optional[Dict[int, Tuple[float, float]]], default=None
-            A dictionary of node positions, where keys are node indices and values are (x, y) coordinates.
-        title : str, default="Graph"
-            The title of the graph plot.
-        node_size : int, default=300
-            The size of the nodes in the plot.
-        with_labels : bool, default=True
-            Whether or not to display labels on the nodes.
-
-        Output:
-        -------
-        Figure
-            The matplotlib figure object representing the final graph plot.
-
-        Description:
-        -----------
-            This function plots a colored graph where each node's color represents its label.
-        """
-        node_labels: Dict[int, Any] = nx.get_node_attributes(G, 'label')
-        
-        unique_labels: Set[Any] = set(node_labels.values())
-        
-        colors: List[str] = ['red', 'black']  # Add more colors if you have more labels
-        
-        color_map: Dict[Any, str] = {label: colors[i % len(colors)] for i, label in enumerate(unique_labels)}
-        
-        node_color: List[str] = [color_map[node_labels[node]] for node in G.nodes]
-        
-        plt.figure(figsize=(10, 8))
-        nx.draw(G, pos, with_labels=with_labels, node_size=node_size, node_color=node_color, 
-                edge_color='gray', font_size=10)
-        plt.title(title)
-        return plt.gcf()
-
-
-    def plot_degree_histogram(self, G: nx.Graph) -> Figure:
-        """
-        Input:
-        ----------
-        G : nx.Graph
-            The NetworkX graph whose degree distribution will be plotted.
-
-        Output:
-        -------
-        Figure
-            The matplotlib figure object representing the histogram plot.
-
-        Description:
-        -----------
-            This function plots a histogram of the node degree frequency distribution for a graph.
-
-        """
-        
-        degrees: List[int] = [val for (_, val) in G.degree()]
-        
-        # Get the unique degrees and their frequencies
-        degree_counts: np.ndarray = np.bincount(degrees)
-        degrees_unique: np.ndarray = np.arange(len(degree_counts))
-        
-        # Normalize the frequencies for colormap
-        max_count: int = max(degree_counts)
-        normalized_counts: np.ndarray = np.array(degree_counts) / max_count if max_count > 0 else degree_counts
-        colors = cm.Blues(normalized_counts)
-        
-        fig, ax = plt.subplots(figsize=(12, 6))
-        ax.bar(degrees_unique, degree_counts, color=colors, align='center')
-        
-        ax.set_xlabel('Degree')
-        ax.set_ylabel('Frequency')
-        ax.set_title('Degree Distribution Histogram')
-        ax.set_xlim(0, 10)
-        
-        return fig
-
-
-    def calculate_common_neighbors(self, edge_index: torch.Tensor, A: 'scipy.sparse.csr_matrix', batch_size: int = 10000) -> torch.FloatTensor:
-        """
-        Input:
-        - edge_index (torch.Tensor): A tensor of shape (2, E), where E is the number of edges. 
-        - A ('scipy.sparse.csr_matrix'): The adjacency matrix of the graph in CSR (Compressed Sparse Row) format.
-        - batch_size (int): The batch size for processing pairs of nodes. Default is 10000.
-
-        Output:
-        - torch.FloatTensor: A tensor containing the number of common neighbors for each pair of nodes in `edge_index`.
-
-        Description:
-            This function computes the number of common neighbors for each pair of nodes in `edge_index` using the adjacency matrix `A`.
-        """
-        link_loader = torch.utils.data.DataLoader(range(edge_index.size(1)), batch_size)
-        scores = []
-
-        for ind in link_loader:
-            src, dst = edge_index[0, ind].to('cpu'), edge_index[1, ind].to('cpu')
-            cur_scores = np.array(A[src].multiply(A[dst]).sum(axis=1)).flatten()
-            scores.append(cur_scores)
-
-        scores = np.concatenate(scores, axis=0)
-        return torch.FloatTensor(scores)
-
-
-    def plot_common_neighbors_distribution(self, train_cn: torch.FloatTensor, test_cn: torch.FloatTensor) -> None:
-        """
-        Input:
-        - train_cn (torch.FloatTensor): A tensor containing the number of common neighbors for the training dataset.
-        - test_cn (torch.FloatTensor): A tensor containing the number of common neighbors for the testing dataset.
-
-        Output:
-        - None: The function saves the plot of the distribution as a PNG file.
-
-        Description:
-        This function generates a density plot of the number of common neighbors for both the training and testing datasets
-        using Kernel Density Estimation (KDE) with the Seaborn library. It visualizes the distribution and saves the plot
-        to a specified file path.
-        """
-
-        # Convert tensors to NumPy arrays for plotting
-        train_cn = train_cn.to('cpu').numpy()
-        test_cn = test_cn.to('cpu').numpy()
-        
-        # print(train_cn)
-        plt.figure(figsize=(10, 6))
-        sns.kdeplot(train_cn, label='Training', color='red', linewidth=2)
-        sns.kdeplot(test_cn, label='Testing', color='blue', linewidth=2)
-        plt.xlabel('Number of Common Neighbors')
-        plt.ylabel('Density')
-        plt.legend()
-        plt.title('Distribution of Common Neighbors')
-        plt.savefig(f'{SyntheticGraphGeneration.FILE_PATH}/common_neighbours_distribution.png')
-
-
-    def generate_nodefeats(self, G: nx.Graph) -> torch.Tensor:
-        """
-        Input:
-        ----------
-        G : nx.Graph
-            The NetworkX graph for which node features will be generated.
-
-        Output:
-        -------
-        torch.Tensor
-            A tensor containing the generated node features.
-
-        Description:
-        -----------
-            This function generates node features for the graph based on the specified feature type.
-            The features can be 'random', 'one-hot', based on the node 'degree'.
-        """
-        num_nodes: int = len(G.nodes)
-        if self.feature_type == 'random':
-            nodefeats: torch.Tensor = torch.randn(num_nodes, self.emb_dim)
-        elif self.feature_type == 'one-hot':
-            nodefeats: torch.Tensor = torch.eye(num_nodes)
-        elif self.feature_type == 'degree':
-            degree: List[int] = [val for (_, val) in G.degree()]
-            nodefeats: torch.Tensor = torch.tensor(degree, dtype=torch.float32).view(-1, 1)
-            
-        return nodefeats
-
-
     def rename_fields(self, data: Data) -> Data:
         """
         Input:
@@ -263,6 +61,12 @@ class SyntheticGraphGeneration:
         
         return data
 
+
+    def create_grid_graph(m, n):
+        """ Create a grid graph and return its NetworkX graph and positions. """
+        G = nx.grid_2d_graph(m, n)
+        pos = {(x, y): (x, y) for x, y in G.nodes()}
+        return G, pos
 
     def create_square_grid(self) -> Tuple[nx.Graph, Dict[int, Tuple[float, float]]]:
         """
